requirements.txt
fastapi
uvicorn
pydantic
mysql-connector-python
pydantic[email]